REF: misplaced Series.combine_first tests
